requirements.txt
asgiref==3.5.2
certifi==2023.11.17
distlib==0.3.8
Django==4.1.13
django-rest-framework==0.1.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.3.1
filelock==3.13.1
pipenv==2023.11.15
platformdirs==4.1.0
psycopg2==2.9.9
PyJWT==2.8.0
pytz==2022.1
setuptools==69.0.2
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.25.0
